requirements.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
jpy==0.11.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
llvmlite==0.38.1
Markdown==3.4.1
MarkupSafe==2.1.1
numba==0.55.2
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pkg_resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slack-sdk==3.18.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.2.0
urllib3==1.26.11
websocket-client==1.3.3
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.0